LayoutTests/media/modern-media-controls/layout-node/layout-node-parent-of-type-expected.txt

Testing the LayoutNode.parentOfType method.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS childNode.parentOfType(CustomNode) is rootNode
PASS childNode.parentOfType(OtherCustomNode) is null

PASS successfullyParsed is true

TEST COMPLETE
